EDU-547 Get rid of answer extension for Task Files
Create special invisible excluded directory to perform all the replacement inside it

+++ python/educational-core/course-creator/src/com/jetbrains/edu/coursecreator/CCFileDeletedListener.java
@@ -25,6 +25,9 @@ class CCFileDeletedListener extends VirtualFileAdapter {
       return;
     }
     VirtualFile removedFile = event.getFile();
+    if (removedFile.getPath().contains(CCUtils.GENERATED_FILES_FOLDER)) {
+      return;
+    }
     final TaskFile taskFile = CCProjectService.getInstance(myProject).getTaskFile(removedFile);
     if (taskFile != null) {
       deleteAnswerFile(removedFile, taskFile);
